condense sheltering to one file
